+++ utils/__init__.py
@@ -1,1 +1,2 @@
 from .dataloader import *
+from .visual import *
